requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
matplotlib>=3.7

sentence-transformers>=2.6.0
torch>=2.2.0
tqdm>=4.66

matplotlib>=3.8

umap-learn
matplotlib